Math: make ComplexNumber mutable
For large screen size and deep levels of zoom, the fractal step function
can be called easily 100'000'000, each returning a ComplexNumber and
performing a square and addition, so 200'000'000 very short lived instances
of ComplexNumber are created. This creates huge pressure on the finalizer
and the garbage collector. Making it mutable saves most of this pressure
(2 instances per pixel are still created, z0 and zn) and will increase
performance at high zoom levels a lot (~40%).

StepFractalFunction, Mandelbrot and all test updated.

Refs #33

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>