Implement Julia fractal
This implements the Julia fractal. The GUI allows setting the constant by
using two fields, one for the real the other for the imaginary part. These
fields are only displayed if the fractal function needs a constant. The
Juli fractal comes pre-defined with a constant of "-0.4+0.6i" because this
displays a nice image.

Fixes the Scaler as well as it had the y-axis mirrored.

Refs #37, #40

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>