CycleFinder: try to find all cycles by scanning the whole boundary area
This tries to find all cycles by scanning the whole boundary area for start
points that lead to a cycle not found previously. The boundary is divided
into (5 * cycleLength) parts in both directions.

This could still be optimized: ~9 objects are instantiated for each of
the starting points, but it already runs reasonably fast.

And the GUI really starts to get crowded.

And documentation is still missing. My favourite part.

Refs #45

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>