Orbits: draw a little circle around orbit stating point
With cyclic orbits, it's not possible to tell where the cycle starts.
Also with Julia fractals, each corner of a cyclic orbit is a starting point.
This is confusing since the color of a cycle changes several times, making
one think the cycle finder finds the same starting point many times when
in fact it doesn't. To "fix" this confusion, draw a little circle, 10 pixels
in diameter, around each starting point, giving a nice visual effect.

Refs #45

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>