Avoid resizing of the info panel
On OSX the fonts are different, more space is needed. This leads to resizing
of the info panel and because of this also a resize of the display area.
This in turn gets the fractal redrawn. The resizing comes from the chaning
coordinate on mouse-over. Fix it by filling the imaginary part label with
the longest number possible, use that as preferred size for the visible
label. This way it works on all platforms (hopefully).

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>