Fix deadlock in fractal redraw
When it happens that a previous redraw finishes just the exact same moment
a new draw request occurs, it deadlocks on the ComplexEscapeFractal.
The drawFractal() methods owns the monitor for ComplexEscapeFractal, waits
for the drawing thread via thread.join(). The drawing thread tries to
acquire the monitor for ComplexEscapeFractal to set "thread = null"
=> DEADLOCK. Fix it by not trying to set the thread to null.
Thread.interrupt() and Thread.join() will do the right thing if the thread
has already exitted.

Fixes #60

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>