Minor GUI fixes for OSX: avoid filckering for Julia
The flickering avoidance is not effecive enough when using Julia fractals.
Change/resize some things to make it work. And force the minimum size.
Also, const real/imag on the same line to save some space. Those don't
need to be as long as the orbit real/imag. Next step is to unclutter
the cycle finder a bit, maybe.

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>