Experimental: Fractal drawing in two threads
Draw the fractal in two threads: the first does coarse to fine except for
the last and finest step. This one takes the most time. Instead the fine
step is calculated in a separate thread, started before the coarse to fine
calculations begin. This is not an optimal usage of two CPUs, but it's
simple enough and gives an improvement. Dunno if it's worth it...

Refs #42

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>