Restructure GUI, option to only draw orbit cycle start point
Restructure the GUI, mainly to make room for the checkbox in the cycle finder.
As as a result, the GUI should be a bit more intuitive to use as well.
The new checkbox, "Full orbit" controls if the full orbit is drawn for
every cycle found. When unchecked, the steps of the orbit to draw is simply
set to 0.

Also, since the "Draw" and the "Reset" button are no longer on the right
side, there's room for the upcoming "save" button.

Fixes #63

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>