Show progress for cycle finder
Shows the progress of the cycle finder. The progress is calculated from the
starting points that are tried vs. those done. The progress bar is only shown
during the search, replaced by the number of cycles found when done. To be
consistent, this also changes the fractal progress bar to use the same
behavior.
Also, this changes the way the starting points are determined. Instead of
5*cycleLength steps in each direction, this uses cycleLength^2 steps in
each direction to find more cycles.

Fixes #64.

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>