Orbit drawing: draw directly on layer to fix thick lines
Animated orbit drawing caused lines to get thicker with each step of the
animation due to antialiasing and the fact that the same images was drawn
again and again (30x) onto the layer. Since orbit drawing is a single thread
now, it's possible to draw directly onto the layer. This fixes the thick
lines, saves memory and increases speed while to code is simpler. Also,
this makes it possible to use a thicker stroke (0.7 instead of 0.4) to draw
all lines, increasing visibility. Extends the testcase for it.

Since it is faster, the cycle finder minimum delay is now half, 50ms.

Fixes #68

Signed-off-by: Daniel Ritz <daniel.ritz@gmx.ch>